+++ src/Core/SoftLine.Trebol.Application/Features/Third/Queries/Vms/ThirdPartyVm.cs
@@ -2,10 +2,10 @@
 {
     public class ThirdPartyVm
     {
-        public decimal NIT { get; set; }
+        public int NIT { get; set; }
         public string? Class { get; set; }
         public string? Regime { get; set; }
-        public decimal? VerificationDigitNIT { get; set; }
+        public int? VerificationDigitNIT { get; set; }
         public string? NITCode { get; set; }
         public string? BusinessName { get; set; }
         public string? Address { get; set; }
